[REF] html_editor, *: move custom snippet methods to website
*: html_builder, website

Following website refactoring [1], and web_editor deletion, models were
moved to html_editor [2].
This commit is moving `ir_ui_view` and `SnippetModel` methods related to
`save_snippet` to website, as they are specific to the module.

The original method definitions on `SnippetModel` were kept, but their
logic was removed, providing an empty structure for other modules to
extend in the future.

[1]: https://github.com/odoo/odoo/commit/9fe45e2b7ddbbfd0445ffe25a859e67a316d02b2
[2]: https://github.com/odoo/odoo/commit/867d9a2f923a55c0bd6a281e2dea1f524fa04445

task-5065980
