+++ addons/html_editor/tests/test_views.py
@@ -61,16 +61,3 @@ class TestViews(TransactionCase):
             '<p>Hello World!</p>',
             base.inherit_children_ids.get_combined_arch(),
         )
-
-    def test_find_available_name(self):
-        View = self.env['ir.ui.view']
-        used_names = ['Unrelated name']
-        initial_name = "Test name"
-        name = View._find_available_name(initial_name, used_names)
-        self.assertEqual(initial_name, name)
-        used_names.append(name)
-        name = View._find_available_name(initial_name, used_names)
-        self.assertEqual('Test name (2)', name)
-        used_names.append(name)
-        name = View._find_available_name(initial_name, used_names)
-        self.assertEqual('Test name (3)', name)
