+++ addons/website/tests/test_views.py
@@ -192,6 +192,19 @@ class TestCustomizeView(common.HttpCase):
         self.assertEqual(set(actives), {'website.test_view'})
         self.assertEqual([custo.active, default.active], [True, True])
 
+    def test_find_available_name(self):
+        View = self.env['ir.ui.view']
+        used_names = ['Unrelated name']
+        initial_name = "Test name"
+        name = View._find_available_name(initial_name, used_names)
+        self.assertEqual(initial_name, name)
+        used_names.append(name)
+        name = View._find_available_name(initial_name, used_names)
+        self.assertEqual('Test name (2)', name)
+        used_names.append(name)
+        name = View._find_available_name(initial_name, used_names)
+        self.assertEqual('Test name (3)', name)
+
 
 class TestViewSaving(TestViewSavingCommon):
